+++ src/CredentialSource/ExecutableSource.php
@@ -88,12 +88,12 @@ class ExecutableSource implements ExternalAccountCredentialSourceInterface
 
     /**
      * @param string $command    The string command to run to get the subject token.
-     * @param string $outputFile
+     * @param string|null $outputFile
      */
     public function __construct(
         string $command,
         ?string $outputFile,
-        ExecutableHandler $executableHandler = null,
+        ?ExecutableHandler $executableHandler = null,
     ) {
         $this->command = $command;
         $this->outputFile = $outputFile;
@@ -101,12 +101,24 @@ class ExecutableSource implements ExternalAccountCredentialSourceInterface
     }
 
     /**
-     * @param callable $httpHandler unused.
+     * Gets the unique key for caching
+     * The format for the cache key is:
+     * Command.OutputFile
+     *
+     * @return ?string
+     */
+    public function getCacheKey(): ?string
+    {
+        return $this->command . '.' . $this->outputFile;
+    }
+
+    /**
+     * @param callable|null $httpHandler unused.
      * @return string
      * @throws RuntimeException if the executable is not allowed to run.
      * @throws ExecutableResponseError if the executable response is invalid.
      */
-    public function fetchSubjectToken(callable $httpHandler = null): string
+    public function fetchSubjectToken(?callable $httpHandler = null): string
     {
         // Check if the executable is allowed to run.
         if (getenv(self::GOOGLE_EXTERNAL_ACCOUNT_ALLOW_EXECUTABLES) !== '1') {