+++ src/OAuth2.php
@@ -581,12 +581,12 @@ class OAuth2 implements FetchAuthTokenInterface
     /**
      * Generates a request for token credentials.
      *
-     * @param callable $httpHandler callback which delivers psr7 request
+     * @param callable|null $httpHandler callback which delivers psr7 request
      * @param array<mixed> $headers [optional] Additional headers to pass to
      *        the token endpoint request.
      * @return RequestInterface the authorization Url.
      */
-    public function generateCredentialsRequest(callable $httpHandler = null, $headers = [])
+    public function generateCredentialsRequest(?callable $httpHandler = null, array $headers = [])
     {
         $uri = $this->getTokenCredentialUri();
         if (is_null($uri)) {
@@ -611,6 +611,9 @@ class OAuth2 implements FetchAuthTokenInterface
                 break;
             case 'refresh_token':
                 $params['refresh_token'] = $this->getRefreshToken();
+                if (isset($this->getAdditionalClaims()['target_audience'])) {
+                    $params['target_audience'] = $this->getAdditionalClaims()['target_audience'];
+                }
                 $this->addClientCredentials($params);
                 break;
             case self::JWT_URN:
@@ -661,12 +664,12 @@ class OAuth2 implements FetchAuthTokenInterface
     /**
      * Fetches the auth tokens based on the current state.
      *
-     * @param callable $httpHandler callback which delivers psr7 request
+     * @param callable|null $httpHandler callback which delivers psr7 request
      * @param array<mixed> $headers [optional] If present, add these headers to the token
      *        endpoint request.
      * @return array<mixed> the response
      */
-    public function fetchAuthToken(callable $httpHandler = null, $headers = [])
+    public function fetchAuthToken(?callable $httpHandler = null, array $headers = [])
     {
         if (is_null($httpHandler)) {
             $httpHandler = HttpHandlerFactory::build(HttpClientCache::getHttpClient());
@@ -683,6 +686,8 @@ class OAuth2 implements FetchAuthTokenInterface
     }
 
     /**
+     * @deprecated
+     *
      * Obtains a key that can used to cache the results of #fetchAuthToken.
      *
      * The key is derived from the scopes.
@@ -704,6 +709,16 @@ class OAuth2 implements FetchAuthTokenInterface
     }
 
     /**
+     * Gets this instance's SubjectTokenFetcher
+     *
+     * @return null|ExternalAccountCredentialSourceInterface
+     */
+    public function getSubjectTokenFetcher(): ?ExternalAccountCredentialSourceInterface
+    {
+        return $this->subjectTokenFetcher;
+    }
+
+    /**
      * Parses the fetched tokens.
      *
      * @param ResponseInterface $resp the response.
@@ -712,7 +727,7 @@ class OAuth2 implements FetchAuthTokenInterface
      */
     public function parseTokenResponse(ResponseInterface $resp)
     {
-        $body = (string)$resp->getBody();
+        $body = (string) $resp->getBody();
         if ($resp->hasHeader('Content-Type') &&
             $resp->getHeaderLine('Content-Type') == 'application/x-www-form-urlencoded'
         ) {
@@ -997,13 +1012,13 @@ class OAuth2 implements FetchAuthTokenInterface
         if (!$this->isAbsoluteUri($uri)) {
             // "postmessage" is a reserved URI string in Google-land
             // @see https://developers.google.com/identity/sign-in/web/server-side-flow
-            if ('postmessage' !== (string)$uri) {
+            if ('postmessage' !== (string) $uri) {
                 throw new InvalidArgumentException(
                     'Redirect URI must be absolute'
                 );
             }
         }
-        $this->redirectUri = (string)$uri;
+        $this->redirectUri = (string) $uri;
     }
 
     /**
@@ -1018,6 +1033,16 @@ class OAuth2 implements FetchAuthTokenInterface
         }
 
         return implode(' ', $this->scope);
+    }
+
+    /**
+     * Gets the subject token type
+     *
+     * @return ?string
+     */
+    public function getSubjectTokenType(): ?string
+    {
+        return $this->subjectTokenType;
     }
 
     /**
@@ -1105,7 +1130,7 @@ class OAuth2 implements FetchAuthTokenInterface
                     'invalid grant type'
                 );
             }
-            $this->grantType = (string)$grantType;
+            $this->grantType = (string) $grantType;
         }
     }
 
@@ -1438,7 +1463,7 @@ class OAuth2 implements FetchAuthTokenInterface
             $this->issuedAt = null;
         } else {
             $this->issuedAt = time();
-            $this->expiresIn = (int)$expiresIn;
+            $this->expiresIn = (int) $expiresIn;
         }
     }
 
@@ -1661,11 +1686,11 @@ class OAuth2 implements FetchAuthTokenInterface
      *
      * Alias of {@see Google\Auth\OAuth2::getClientId()}.
      *
-     * @param callable $httpHandler
+     * @param callable|null $httpHandler
      * @return string
      * @access private
      */
-    public function getClientName(callable $httpHandler = null)
+    public function getClientName(?callable $httpHandler = null)
     {
         return $this->getClientId();
     }
@@ -1746,7 +1771,8 @@ class OAuth2 implements FetchAuthTokenInterface
                 throw new \InvalidArgumentException(
                     'To have multiple allowed algorithms, You must provide an'
                     . ' array of Firebase\JWT\Key objects.'
-                    . ' See https://github.com/firebase/php-jwt for more information.');
+                    . ' See https://github.com/firebase/php-jwt for more information.'
+                );
             }
             $allowedAlg = array_pop($allowedAlgs);
         } else {