+++ tests/ApplicationDefaultCredentialsTest.php
@@ -21,15 +21,21 @@ use DomainException;
 use Google\Auth\ApplicationDefaultCredentials;
 use Google\Auth\Credentials\ExternalAccountCredentials;
 use Google\Auth\Credentials\GCECredentials;
+use Google\Auth\Credentials\ImpersonatedServiceAccountCredentials;
 use Google\Auth\Credentials\ServiceAccountCredentials;
+use Google\Auth\Credentials\UserRefreshCredentials;
 use Google\Auth\CredentialsLoader;
 use Google\Auth\CredentialSource;
+use Google\Auth\FetchAuthTokenCache;
 use Google\Auth\GCECache;
+use Google\Auth\Logging\StdOutLogger;
 use GuzzleHttp\Psr7;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Psr7\Utils;
+use PHPUnit\Framework\Error\Notice;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
+use Psr\Cache\CacheItemPoolInterface;
 use ReflectionClass;
 
 /**
@@ -43,6 +49,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     private $targetAudience = 'a target audience';
     private $quotaProject = 'a-quota-project';
     private $originalServiceAccount;
+    private const SDK_DEBUG_ENV_VAR = 'GOOGLE_SDK_PHP_LOGGING';
 
     public function testGetCredentialsFailsIfEnvSpecifiesNonExistentFile()
     {
@@ -103,7 +110,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
         ]);
 
         $this->assertInstanceOf(
-            'Google\Auth\Credentials\GCECredentials',
+            GCECredentials::class,
             ApplicationDefaultCredentials::getCredentials('a scope', $httpHandler)
         );
     }
@@ -126,10 +133,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             'a+default+scope' // $defaultScope
         );
 
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\GCECredentials',
-            $creds
-        );
+        $this->assertInstanceOf(GCECredentials::class, $creds);
 
         $uriProperty = (new ReflectionClass($creds))->getProperty('tokenUri');
         $uriProperty->setAccessible(true);
@@ -166,10 +170,8 @@ class ApplicationDefaultCredentialsTest extends TestCase
             null,
             'a default scope'
         );
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\ImpersonatedServiceAccountCredentials',
-            $creds);
 
+        $this->assertInstanceOf(ImpersonatedServiceAccountCredentials::class, $creds);
         $this->assertEquals('service_account_name@namespace.iam.gserviceaccount.com', $creds->getClientName());
 
         $sourceCredentialsProperty = (new ReflectionClass($creds))->getProperty('sourceCredentials');
@@ -177,9 +179,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
         // used default scope
         $sourceCredentials = $sourceCredentialsProperty->getValue($creds);
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\UserRefreshCredentials',
-            $sourceCredentials);
+        $this->assertInstanceOf(UserRefreshCredentials::class, $sourceCredentials);
     }
 
     public function testUserRefreshCredentials()
@@ -195,10 +195,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             'a default scope' // $defaultScope
         );
 
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\UserRefreshCredentials',
-            $creds
-        );
+        $this->assertInstanceOf(UserRefreshCredentials::class, $creds);
 
         $authProperty = (new ReflectionClass($creds))->getProperty('auth');
         $authProperty->setAccessible(true);
@@ -234,10 +231,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             'a default scope' // $defaultScope
         );
 
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\ServiceAccountCredentials',
-            $creds
-        );
+        $this->assertInstanceOf(ServiceAccountCredentials::class, $creds);
 
         $authProperty = (new ReflectionClass($creds))->getProperty('auth');
         $authProperty->setAccessible(true);
@@ -329,7 +323,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
         ]);
 
         $cacheOptions = [];
-        $cachePool = $this->prophesize('Psr\Cache\CacheItemPoolInterface');
+        $cachePool = $this->prophesize(CacheItemPoolInterface::class);
 
         $middleware = ApplicationDefaultCredentials::getMiddleware(
             'a scope',
@@ -372,7 +366,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             ->shouldBeCalledTimes(1)
             ->willReturn(false);
 
-        $mockCache = $this->prophesize('Psr\Cache\CacheItemPoolInterface');
+        $mockCache = $this->prophesize(CacheItemPoolInterface::class);
         $mockCache->getItem(GCECache::GCE_CACHE_KEY)
             ->shouldBeCalledTimes(1)
             ->willReturn($mockCacheItem->reveal());
@@ -404,7 +398,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             ->shouldBeCalledTimes(1)
             ->willReturn($mockCacheItem->reveal());
 
-        $mockCache = $this->prophesize('Psr\Cache\CacheItemPoolInterface');
+        $mockCache = $this->prophesize(CacheItemPoolInterface::class);
         $mockCache->getItem(GCECache::GCE_CACHE_KEY)
             ->shouldBeCalledTimes(2)
             ->willReturn($mockCacheItem->reveal());
@@ -443,7 +437,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             ->shouldBeCalledTimes(1)
             ->willReturn($mockCacheItem->reveal());
 
-        $mockCache = $this->prophesize('Psr\Cache\CacheItemPoolInterface');
+        $mockCache = $this->prophesize(CacheItemPoolInterface::class);
         $mockCache->getItem($prefix . GCECache::GCE_CACHE_KEY)
             ->shouldBeCalledTimes(2)
             ->willReturn($mockCacheItem->reveal());
@@ -475,20 +469,20 @@ class ApplicationDefaultCredentialsTest extends TestCase
         putenv(ServiceAccountCredentials::ENV_VAR . '=' . $keyFile);
 
         $creds = ApplicationDefaultCredentials::getIdTokenCredentials($this->targetAudience);
-
-        $this->assertNotNull($creds);
+        $this->assertInstanceOf(ServiceAccountCredentials::class, $creds);
     }
 
     public function testGetIdTokenCredentialsLoadsDefaultFileIfPresentAndEnvVarIsNotSet()
     {
         putenv('HOME=' . __DIR__ . '/fixtures');
         $creds = ApplicationDefaultCredentials::getIdTokenCredentials($this->targetAudience);
-        $this->assertNotNull($creds);
+        $this->assertInstanceOf(ServiceAccountCredentials::class, $creds);
     }
 
     public function testGetIdTokenCredentialsFailsIfNotOnGceAndNoDefaultFileFound()
     {
         $this->expectException(DomainException::class);
+        $this->expectExceptionMessage('Your default credentials were not found');
 
         putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
 
@@ -499,12 +493,10 @@ class ApplicationDefaultCredentialsTest extends TestCase
             new Response(500)
         ]);
 
-        $creds = ApplicationDefaultCredentials::getIdTokenCredentials(
+        ApplicationDefaultCredentials::getIdTokenCredentials(
             $this->targetAudience,
             $httpHandler
         );
-
-        $this->assertNotNull($creds);
     }
 
     public function testGetIdTokenCredentialsWithCacheOptions()
@@ -517,7 +509,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
         ]);
 
         $cacheOptions = [];
-        $cachePool = $this->prophesize('Psr\Cache\CacheItemPoolInterface');
+        $cachePool = $this->prophesize(CacheItemPoolInterface::class);
 
         $credentials = ApplicationDefaultCredentials::getIdTokenCredentials(
             $this->targetAudience,
@@ -526,7 +518,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             $cachePool->reveal()
         );
 
-        $this->assertInstanceOf('Google\Auth\FetchAuthTokenCache', $credentials);
+        $this->assertInstanceOf(FetchAuthTokenCache::class, $credentials);
     }
 
     public function testGetIdTokenCredentialsSuccedsIfNoDefaultFilesButIsOnGCE()
@@ -550,10 +542,27 @@ class ApplicationDefaultCredentialsTest extends TestCase
             $httpHandler
         );
 
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\GCECredentials',
-            $credentials
+        $this->assertInstanceOf(GCECredentials::class, $credentials);
+    }
+
+    public function testGetIdTokenCredentialsWithUserRefreshCredentials()
+    {
+        putenv('HOME=' . __DIR__ . '/fixtures2');
+
+        $creds = ApplicationDefaultCredentials::getIdTokenCredentials(
+            $this->targetAudience,
         );
+
+        $this->assertInstanceOf(UserRefreshCredentials::class, $creds);
+
+        $authProperty = (new ReflectionClass($creds))->getProperty('auth');
+        $authProperty->setAccessible(true);
+
+        // used default scope
+        $auth = $authProperty->getValue($creds);
+        $additionalClaims = $auth->getAdditionalClaims();
+        $this->assertArrayHasKey('target_audience', $additionalClaims);
+        $this->assertEquals($this->targetAudience, $additionalClaims['target_audience']);
     }
 
     public function testWithServiceAccountCredentialsAndExplicitQuotaProject()
@@ -569,10 +578,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             $this->quotaProject
         );
 
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\ServiceAccountCredentials',
-            $credentials
-        );
+        $this->assertInstanceOf(ServiceAccountCredentials::class, $credentials);
 
         $this->assertEquals(
             $this->quotaProject,
@@ -656,7 +662,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
         ]);
 
         $cacheOptions = [];
-        $cachePool = $this->prophesize('Psr\Cache\CacheItemPoolInterface');
+        $cachePool = $this->prophesize(CacheItemPoolInterface::class);
 
         $credentials = ApplicationDefaultCredentials::getCredentials(
             null,
@@ -666,7 +672,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             $this->quotaProject
         );
 
-        $this->assertInstanceOf('Google\Auth\FetchAuthTokenCache', $credentials);
+        $this->assertInstanceOf(FetchAuthTokenCache::class, $credentials);
 
         $this->assertEquals(
             $this->quotaProject,
@@ -698,10 +704,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             $this->quotaProject
         );
 
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\GCECredentials',
-            $credentials
-        );
+        $this->assertInstanceOf(GCECredentials::class, $credentials);
 
         $this->assertEquals(
             $this->quotaProject,
@@ -728,7 +731,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             new Response(200, [GCECredentials::FLAVOR_HEADER => 'Google']),
         ]);
         $this->assertInstanceOf(
-            'Google\Auth\Credentials\GCECredentials',
+            GCECredentials::class,
             ApplicationDefaultCredentials::getCredentials(null, $httpHandler)
         );
     }
@@ -745,10 +748,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
             $this->targetAudience,
             $httpHandler
         );
-        $this->assertInstanceOf(
-            'Google\Auth\Credentials\GCECredentials',
-            $creds
-        );
+        $this->assertInstanceOf(GCECredentials::class, $creds);
     }
 
     /**
@@ -773,6 +773,40 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
         $subjectTokenFetcher = $oauthProp->getValue($oauth);
         $this->assertInstanceOf($expectedCredSource, $subjectTokenFetcher);
+    }
+
+    public function testGetDefaultLoggerReturnStdOutLoggerIfEnvVarIsPresent()
+    {
+        putenv($this::SDK_DEBUG_ENV_VAR . '=true');
+        $logger = ApplicationDefaultCredentials::getDefaultLogger();
+        $this->assertTrue($logger instanceof StdOutLogger);
+    }
+
+    public function testGetDefaultLoggerReturnsNullIfNotEnvVar()
+    {
+        putenv($this::SDK_DEBUG_ENV_VAR . '=false');
+        $logger = ApplicationDefaultCredentials::getDefaultLogger();
+
+        $this->assertNull($logger);
+
+        putenv($this::SDK_DEBUG_ENV_VAR . '=0');
+        $logger = ApplicationDefaultCredentials::getDefaultLogger();
+
+        $this->assertNull($logger);
+
+        putenv($this::SDK_DEBUG_ENV_VAR . '=');
+        $logger = ApplicationDefaultCredentials::getDefaultLogger();
+
+        $this->assertNull($logger);
+    }
+
+    public function testGetDefaultLoggerRaiseAWarningIfMisconfiguredAndReturnsNull()
+    {
+        putenv($this::SDK_DEBUG_ENV_VAR . '=invalid');
+        $this->expectException(Notice::class);
+        $logger = ApplicationDefaultCredentials::getDefaultLogger();
+
+        $this->assertNull($logger);
     }
 
     public function provideExternalAccountCredentials()