+++ tests/CredentialSource/FileSourceTest.php
@@ -32,8 +32,8 @@ class FileSourceTest extends TestCase
     public function testFetchSubjectToken(
         string $filename,
         string $expectedToken,
-        string $format = null,
-        string $subjectTokenFieldName = null
+        ?string $format = null,
+        ?string $subjectTokenFieldName = null
     ) {
         $source = new FileSource($filename, $format, $subjectTokenFieldName);
         $subjectToken = $source->fetchSubjectToken();
@@ -44,7 +44,6 @@ class FileSourceTest extends TestCase
     {
         $file1 = tempnam(sys_get_temp_dir(), 'test1');
         file_put_contents($file1, 'abc');
-
 
         $file2 = tempnam(sys_get_temp_dir(), 'test2');
         file_put_contents($file2, json_encode(['token' => 'def']));